Refactor uri matcher
